+++ types/node/node-tests.ts
@@ -1,5 +1,4 @@
 import * as assert from 'assert';
-import { promisify } from "util";
 
 /**
  * The `expectType` function from https://www.npmjs.com/package/tsd,
@@ -61,20 +60,4 @@ assert['fail'](true, true, 'works like a charm');
     const a = { b: 2 } as any;
     assert.deepStrictEqual(a, { b: 2 });
     a; // $ExpectType { b: number; }
-}
-
-// Test cases from https://github.com/microsoft/TypeScript/issues/41563#issuecomment-729270419
-{
-    function parseString(str: any, cb?: (...args: any[]) => any): void {}
-    const parseStringP: (str: any) => Promise<any> = promisify(parseString);
-    parseStringP("abc");
-
-    const arg1UnknownError: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: unknown, result: number) => void): void => { });
-    const arg1AnyError: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: any, result: number) => void): void => { });
-    const arg0: () => Promise<number> = promisify((cb: (err: Error | null, result: number) => void): void => { });
-    const arg0NoResult: () => Promise<void> = promisify((cb: (err: Error | null) => void): void => { });
-    const arg1: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: Error | null, result: number) => void): void => { });
-    const arg1UnknownError2: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: Error | null, result: number) => void): void => { });
-    const arg1NoResult: (arg: string) => Promise<void> = promisify((arg: string, cb: (err: Error | null) => void): void => { });
-    const cbOptionalError: () => Promise<void> = promisify((cb: (err?: Error | null) => void): void => { cb(); });
 }
