+++ types/node/package.json
@@ -11,11 +11,6 @@
             "*": [
                 "ts3.6/*"
             ]
-        },
-        "<=3.9": {
-            "*": [
-                "ts3.9/*"
-            ]
         }
     }
 }
