+++ types/node/ts3.6/index.d.ts
@@ -5,4 +5,3 @@
 /// <reference path="base.d.ts" />
 
 /// <reference path="../ts3.4/assert.d.ts" />
-/// <reference path="../ts3.4/util.d.ts" />
