--- types/node/ts3.9/base.d.ts
@@ -1,7 +1,0 @@
-// NOTE: These definitions support NodeJS and TypeScript up to 3.9.
-
-// Base definitions for all NodeJS modules that are not specific to any version of TypeScript:
-/// <reference path="../ts3.6/base.d.ts" />
-
-// TypeScript 3.7-specific augmentations:
-/// <reference path="../assert.d.ts" />
