--- types/node/ts3.9/index.d.ts
@@ -1,5 +1,0 @@
-// NOTE: These definitions support NodeJS and TypeScript up to 3.9.
-// Newer versions use conditional and variadic tuples to resolve the promisify signature
-
-/// <reference path="base.d.ts" />
-/// <reference path="../ts3.4/util.d.ts" />
