--- types/node/ts3.9/node-tests.ts
@@ -1,1 +1,0 @@
-import '../ts3.6/node-tests';
