--- types/node/ts3.9/tslint.json
@@ -1,6 +1,0 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "no-bad-reference": false
-    }
-}
